requirements.txt
wheel==0.41.1
fastapi==0.101.0
uvicorn==0.23.2
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
asyncpg==0.29.0
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.0.4
openpyxl==3.1.2
numpy==1.25.2
pandas==2.0.3
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9
boto3==1.34.134